+++ src/sign.rs
@@ -40,7 +40,7 @@ impl Signature {
         seed: &[u32; 4],
     ) -> Self {
         let mut rng = ChaChaRng::from_seed(seed);
-        let time_vec = time_to_fr_vec(*time as u32, CONST_D as u32);
+        let time_vec = time_to_fr_vec(*time, CONST_D as u32);
         Self::sign(ssk, pp, &time_vec, msg, &mut rng)
     }
 }
